Weird behaviour when starting new conversation fixed Notes ownership bug fixed Unneeded packages removed Chats are becoming more real-time

+++ app/Http/Controllers/User/CategoryController.php
@@ -17,8 +17,8 @@ class CategoryController extends Controller
         $data = $request->getParsedBody();
 
         return Connection::get()->query(
-            'INSERT INTO categories(name, created_at, updated_at) VALUES (?, ?, ?);',
-            [$data['name'], time(), time()]
+            'INSERT INTO categories(name, user_id, created_at, updated_at) VALUES (?, ?, ?, ?);',
+            [$data['name'], request()->auth()->userId(), time(), time()]
         )->then(function (Result $result) use (&$data) {
             $data['id'] = $result->insertId;
             return response()->json([
@@ -35,24 +35,26 @@ class CategoryController extends Controller
 
     public function list(ServerRequestInterface $request)
     {
-        return Connection::get()->query('SELECT * FROM categories;')->then(function (Result $result) use (&$data) {
-            return response()->json([
-                'status' => true,
-                'data' => $result->rows
-            ]);
-        })->otherwise(function (Throwable $throwable) {
-            return response()->json([
-                'status' => false,
-                'message' => 'Selection failed'
-            ]);
-        });
+        return Connection::get()
+            ->query('SELECT * FROM categories WHERE user_id = ?;', [request()->auth()->userId()])
+            ->then(function (Result $result) use (&$data) {
+                return response()->json([
+                    'status' => true,
+                    'data' => $result->rows
+                ]);
+            })->otherwise(function (Throwable $throwable) {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'Selection failed'
+                ]);
+            });
     }
 
     public function open(ServerRequestInterface $request, array $params)
     {
         return Connection::get()->query(
-            'SELECT * FROM notes WHERE category_id = ?;',
-            [$params['id']]
+            'SELECT * FROM notes WHERE category_id = ? AND user_id = ?;',
+            [$params['id'], request()->auth()->userId()]
         )->then(function (Result $result) use (&$data) {
             return response()->json([
                 'status' => true,
@@ -70,8 +72,8 @@ class CategoryController extends Controller
     {
         $data = $request->getParsedBody();
         return Connection::get()->query(
-            'UPDATE categories SET name = ?, updated_at = ? WHERE id = ?;',
-            [$data['name'], time(), $params['id']]
+            'UPDATE categories SET name = ?, updated_at = ? WHERE id = ? AND user_id = ?;',
+            [$data['name'], time(), $params['id'], request()->auth()->userId()]
         )->then(function (Result $result) use (&$data) {
             return response()->json([
                 'status' => true,
@@ -88,8 +90,8 @@ class CategoryController extends Controller
     public function delete(ServerRequestInterface $request, array $params)
     {
         return Connection::get()->query(
-            'DELETE FROM categories WHERE id = ?;',
-            [$params['id']]
+            'DELETE FROM categories WHERE id = ? AND user_id = ?;',
+            [$params['id'], request()->auth()->userId()]
         )->then(function (Result $result) use (&$data) {
             return response()->json([
                 'status' => true,
